requirements.txt
bandit==1.8.6
black==25.9.0
boolean.py==5.0
boto3==1.40.40
botocore==1.40.40
CacheControl==0.14.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
coverage==7.10.7
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
filelock==3.19.1
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
license-expression==30.4.4
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.1
mypy_extensions==1.1.0
numpy==2.3.3
packageurl-python==0.17.5
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.4.0
pluggy==1.6.0
py-serializable==2.1.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.1.0
s3transfer==0.14.0
six==1.17.0
sortedcontainers==2.4.0
stevedore==5.5.0
toml==0.10.2
tzdata==2025.2
urllib3==2.5.0
